# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Digildrobe API",
        "contact": {
            "email": "support@digildrobe.com"
        },
        "version": "0.1"
    },
    "paths": {
        "/api/v1/login/social": {
            "post": {
                "tags": [
                    "Guest"
                ],
                "description": "Guest login by soicial.",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\AuthController::socialLogin",
                "requestBody": {
                    "description": "Social body login",
                    "required": true,
                    "content": {
                        "form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/SocialAuthen"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Guest"
                ],
                "description": "Guest login by local.",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\AuthController::localLogin",
                "requestBody": {
                    "description": "local body login",
                    "required": true,
                    "content": {
                        "form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/LocalAuthen"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "User logout.",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\AuthController::logout",
                "responses": {
                    "200": {
                        "description": "An example resource"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/brands": {
            "get": {
                "tags": [
                    "Guest"
                ],
                "description": "Get list brands.",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\BrandController::index",
                "parameters": [
                    {
                        "name": "sub_category_id",
                        "in": "query",
                        "description": "sub category id to filter by",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Guest"
                ],
                "description": "Guest register.",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\GuestController::register",
                "requestBody": {
                    "description": "Guest body register",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Register"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/genders": {
            "get": {
                "tags": [
                    "Guest"
                ],
                "description": "Get list genders.",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\GuestController::getGenders",
                "responses": {
                    "200": {
                        "description": "An example resource"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/styles": {
            "get": {
                "tags": [
                    "Guest"
                ],
                "description": "Get list styles.",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\GuestController::getStyles",
                "responses": {
                    "200": {
                        "description": "An example resource"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Guest"
                ],
                "description": "Get list categories.",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\GuestController::getCategories",
                "responses": {
                    "200": {
                        "description": "An example resource"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/items": {
            "get": {
                "tags": [
                    "Guest"
                ],
                "description": "Get list brands.",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\ItemController::index",
                "parameters": [
                    {
                        "name": "model_id",
                        "in": "query",
                        "description": "model id to filter by",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "brand_id",
                        "in": "query",
                        "description": "brand id to filter by",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "color_id",
                        "in": "query",
                        "description": "color id to filter by",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Add user items.",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::addItem",
                "requestBody": {
                    "description": "Items body register",
                    "required": true,
                    "content": {
                        "form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Item"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/mesure-types": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Get list mesure types.",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::getMeasureType",
                "parameters": [
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "gender to filter by",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/measure": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "create measure.",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::postMeasure",
                "requestBody": {
                    "description": "Measure body register",
                    "required": true,
                    "content": {
                        "form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Measure"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/sub-categories": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Get list sub categories.",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::getSubCategories",
                "responses": {
                    "200": {
                        "description": "An example resource"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/models": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Get list models.",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::getModels",
                "parameters": [
                    {
                        "name": "brand",
                        "in": "query",
                        "description": "brand to filter by",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "sub_category",
                        "in": "query",
                        "description": "sub category to filter by",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/colors": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Get list colors",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::getColors",
                "responses": {
                    "200": {
                        "description": "An example resource"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Get profile",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::show",
                "responses": {
                    "200": {
                        "description": "An example resource"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/v1/change-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "User update password.",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::updatePassword",
                "requestBody": {
                    "description": "Password body register",
                    "required": true,
                    "content": {
                        "form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePasswordBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    },
                    "422": {
                        "description": "Invalid data"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ErrorReponse": {
                "properties": {
                    "message": {
                        "description": "message error",
                        "type": "string"
                    },
                    "errors": {
                        "description": "An array errors",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Item": {
                "required": [
                    "quantity",
                    "catalogue_id"
                ],
                "properties": {
                    "quantity": {
                        "description": "quantity of item",
                        "type": "integer"
                    },
                    "catalogue_id": {
                        "description": "catalogue id of item",
                        "type": "integer"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Item"
                }
            },
            "LocalAuthen": {
                "required": [
                    "email",
                    "password",
                    "device_type",
                    "uuid"
                ],
                "properties": {
                    "email": {
                        "description": "Email of user",
                        "type": "string"
                    },
                    "password": {
                        "description": "Password of user",
                        "type": "string"
                    },
                    "device_type": {
                        "description": "Type of devive, 1:Android;2:IOS",
                        "type": "integer",
                        "enum": [
                            1,
                            2
                        ]
                    },
                    "uuid": {
                        "description": "device token",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LocalAuthen"
                }
            },
            "Measure": {
                "required": [
                    "gender",
                    "bshape",
                    "waist",
                    "hips",
                    "neck",
                    "sleeve",
                    "sleeve",
                    "chest",
                    "inseam"
                ],
                "properties": {
                    "gender": {
                        "description": "gender of user",
                        "type": "integer"
                    },
                    "bshape": {
                        "description": "body shape",
                        "type": "integer"
                    },
                    "waist": {
                        "description": "waits value",
                        "type": "integer"
                    },
                    "hips": {
                        "description": "hips value",
                        "type": "integer"
                    },
                    "neck": {
                        "description": "neck value",
                        "type": "integer"
                    },
                    "sleeve": {
                        "description": "sleeve value",
                        "type": "integer"
                    },
                    "chest": {
                        "description": "chest value",
                        "type": "integer"
                    },
                    "inseam": {
                        "description": "inseam value",
                        "type": "integer"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Item"
                }
            },
            "Register": {
                "required": [
                    "email",
                    "first_name",
                    "last_name",
                    "dob",
                    "gender",
                    "picture",
                    "state",
                    "city",
                    "country_id",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "Email of user",
                        "type": "string"
                    },
                    "first_name": {
                        "description": "first name of user",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "last name of user",
                        "type": "string"
                    },
                    "dob": {
                        "description": "birthday of user",
                        "type": "string",
                        "format": "date-time"
                    },
                    "gender": {
                        "description": "gender of user",
                        "type": "integer"
                    },
                    "picture": {
                        "description": "avartar of user",
                        "type": "string",
                        "format": "binary"
                    },
                    "state": {
                        "description": "state of user",
                        "type": "string"
                    },
                    "city": {
                        "description": "city of user",
                        "type": "string"
                    },
                    "country_id": {
                        "description": "country of user",
                        "type": "string"
                    },
                    "password": {
                        "description": "Password of user",
                        "type": "string"
                    },
                    "confirm_password": {
                        "description": "Password confirm of user",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Register"
                }
            },
            "SocialAuthen": {
                "required": [
                    "access_token",
                    "provider",
                    "device_type",
                    "uuid"
                ],
                "properties": {
                    "access_token": {
                        "description": "access token of social",
                        "type": "string"
                    },
                    "provider": {
                        "description": "Type of social",
                        "type": "string",
                        "enum": [
                            "facebook",
                            "google",
                            "github"
                        ]
                    },
                    "device_type": {
                        "description": "Type of devive, 1:Android;2:IOS",
                        "type": "integer",
                        "enum": [
                            1,
                            2
                        ]
                    },
                    "uuid": {
                        "description": "device token",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "SocialAuthen"
                }
            },
            "UpdatePasswordBody": {
                "required": [
                    "password",
                    "password_confirm"
                ],
                "properties": {
                    "password": {
                        "description": "password of user",
                        "type": "string"
                    },
                    "password_confirm": {
                        "description": "Password confirm of user",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LocalAuthen"
                }
            }
        }
    }
}